Adicionado hook useForm no formulario de cadastro de eventos
