+++ package.json
@@ -15,6 +15,7 @@
     "@react-navigation/native": "^6.1.2",
     "moment": "^2.29.4",
     "react": "18.1.0",
+    "react-hook-form": "^7.43.9",
     "react-native": "0.70.6",
     "react-native-calendars": "^1.1292.0",
     "react-native-gesture-handler": "^2.8.0",
